pyproject.toml
[tool.poetry]
name = "trashvision"
version = "0.1.0"
description = ""
authors = ["AbdelrhmanNile <abdelrhmannile@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.11,<3.12"
tensorflow-cpu = "^2.12.0"
opencv-python = "^4.7.0.72"
tensorflow-hub = "^0.13.0"
flask = "^2.3.2"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
